title duties now included for both SubjectHeader and SubjectDownload
