+++ packages/frontend/app/components/reports/subject-header.js
@@ -1,12 +1,35 @@
 import Component from '@glimmer/component';
-import { tracked } from '@glimmer/tracking';
+import { cached, tracked } from '@glimmer/tracking';
 import { action } from '@ember/object';
+import { service } from '@ember/service';
 import { dropTask } from 'ember-concurrency';
 import { validatable, Length } from 'ilios-common/decorators/validation';
+import { TrackedAsyncData } from 'ember-async-data';
 
 @validatable
 export default class ReportsSubjectHeader extends Component {
+  @service reporting;
   @tracked @Length(1, 240) title = '';
+
+  @cached
+  get reportTitleData() {
+    return new TrackedAsyncData(
+      this.reporting.buildReportTitle(
+        this.args.subject,
+        this.args.prepositionalObject,
+        this.args.prepositionalObjectTableRowId,
+        this.args.school,
+      ),
+    );
+  }
+
+  get reportTitle() {
+    if (this.args.report?.title) {
+      return this.args.report.title;
+    }
+
+    return this.reportTitleData.isResolved ? this.reportTitleData.value : null;
+  }
 
   changeTitle = dropTask(async () => {
     this.addErrorDisplayFor('title');
